requirements.txt
requests
pandas
zwift-client

